vtools: bugfix: use PATH to find `vtools` command

+++ tools/bootstrap.sh
@@ -8,40 +8,41 @@ vroot="$(cd "$(dirname "${BASH_SOURCE[0]}")/.." && pwd)"
 
 echo "Root of project: $vroot"
 
-# install OS package deps
-sudo "${vroot}/tools/install-deps.sh"
+if [[ ! -e "${vroot}/build/bin/vtools" ]]; then
+    # install OS package deps
+    sudo "${vroot}/tools/install-deps.sh"
 
-# install tox (via virtualenv)
-pip install --user tox
-PATH=$HOME/.local/bin:$PATH
+    # install tox (via virtualenv)
+    pip install --user tox
+    PATH="${HOME}/.local/bin:${PATH}:${vroot}/build/bin"
 
-# build vtools
-tox -c tools/
-
+    # build vtools
+    tox -c tools/
+fi
 # install go dependencies
-"${vroot}/build/bin/vtools install go-compiler"
-"${vroot}/build/bin/vtools install go-deps"
+vtools install go-compiler
+vtools install go-deps
 
 # install clang
-"${vroot}/build/bin/vtools install clang"
+vtools install clang
 
 # install external dependencies for all build types
-"${vroot}/build/bin/vtools install cpp-deps"
-"${vroot}/build/bin/vtools install cpp-deps --clang"
-"${vroot}/build/bin/vtools install cpp-deps --build-type debug"
-"${vroot}/build/bin/vtools install cpp-deps --build-type debug --clang"
+vtools install cpp-deps
+vtools install cpp-deps --clang
+vtools install cpp-deps --build-type debug
+vtools install cpp-deps --build-type debug --clang
 
 echo ""
 echo "###############################################"
 echo ""
 echo "successfully installed dev dependencies"
 echo ""
-echo "vtools command available as '${vroot}/build/bin/vtools'"
+echo "vtools command available as 'vtools'"
 echo ""
 echo "###############################################"
 echo ""
 
-"${vroot}/build/bin/vtools" --help
+vtools --help
 
 echo ""
 echo "###############################################"
